Implement simple unit test for complete bipartite graph
Test the adjacnecy matrix and the Laplacian matrix. Also, fix adjacency_matrix() error.

+++ hiperwalk/graph/complete_bipartite.py
@@ -126,4 +126,4 @@ class CompleteBipartite(Graph):
         C = np.ones((self._num_vert2, self._num_vert1), dtype=np.int8)
         D = np.zeros((self._num_vert2, self._num_vert2), dtype=np.int8)
         return np.block([[A, B],
-                         [C, D]], dtype=np.int8)
+                         [C, D]])
